[WIP] Board support layer - Tested on bixler and leo

+++ include/mixer.h
@@ -8,6 +8,7 @@ extern "C" {
 #include <stdbool.h>
 
 
+// This enum needs to match the "array_of_mixers" variable in mixer.c
 typedef enum
 {
   QUADCOPTER_PLUS,
@@ -47,10 +48,10 @@ typedef struct
 
 extern command_t _command;
 
-extern int32_t _GPIO_outputs[8];
+extern float _GPIO_outputs[8];
 extern output_type_t _GPIO_output_type[8];
 
-extern int32_t _outputs[8];
+extern float _outputs[8];
 
 void init_PWM();
 void init_mixing();
